+++ warp/_src/fem/cache.py
@@ -418,7 +418,8 @@ the temporary is explicitly detached from the pool using :meth:`detach`.
 The temporary may also be explicitly returned to the pool before destruction using :meth:`release`.
 
 Note: `Temporary` is now a direct alias for `wp.array` with a custom deleter. Convenience `detach` and `release`
-are added at borrow time, as well as a self-pointing `array` attribute is for backward compatibility.
+are added at borrow time. A self-pointing `array` attribute is also added for backward compatibility, but is
+deprecated and will be removed in Warp 1.12.
 """
 
 
@@ -560,22 +561,43 @@ class TemporaryStore:
 
     @staticmethod
     def add_temporary_convenience_methods(temporary: wp.array) -> Temporary:
-        temporary.release = TemporaryStore._release_temporary.__get__(temporary)
-        temporary.detach = TemporaryStore._detach_temporary.__get__(temporary)
-        temporary.array = temporary
+        ref = weakref.ref(temporary)
+        temporary.release = TemporaryStore._release_temporary.__get__(ref)
+        temporary.detach = TemporaryStore._detach_temporary.__get__(ref)
+
+        # Deprecated -- to be removed in 1.12
+        temporary.array = wp.array(
+            ptr=temporary.ptr,
+            capacity=temporary.capacity,
+            shape=temporary.shape,
+            dtype=temporary.dtype,
+            grad=temporary.grad,
+            device=temporary.device,
+            pinned=temporary.pinned,
+            deleter=None,
+        )
+
         return temporary
 
     @staticmethod
-    def _detach_temporary(temporary) -> wp.array:
+    def _detach_temporary(temporary_ref: "weakref.ReferenceType[Temporary]") -> Temporary:
         """Detaches the temporary so it is never returned to the pool"""
+        temporary = temporary_ref()
+        if temporary is None:
+            return None
+
         if temporary.deleter is not None:
             if isinstance(temporary.deleter, TemporaryStore.Pool.Deleter):
                 temporary.deleter.detach(temporary)
         return temporary
 
     @staticmethod
-    def _release_temporary(temporary):
+    def _release_temporary(temporary_ref: "weakref.ReferenceType[Temporary]"):
         """Returns the temporary array to the pool"""
+        temporary = temporary_ref()
+        if temporary is None:
+            return
+
         if temporary.deleter is not None:
             with temporary.device.context_guard:
                 temporary.deleter(temporary.ptr, temporary.capacity)