+++ warp/_src/fem/integrate.py
@@ -1250,6 +1250,18 @@ def _generate_auxiliary_kernels(
     return ((dispatch_kernel, dispatch_tile_size),)
 
 
+def _as_2d_array(array, shape, dtype):
+    return wp.array(
+        data=None,
+        ptr=array.ptr,
+        capacity=array.capacity,
+        device=array.device,
+        shape=shape,
+        dtype=dtype,
+        grad=None if array.grad is None else _as_2d_array(array.grad, shape, dtype),
+    )
+
+
 def _launch_integrate_kernel(
     integrand: Integrand,
     kernel: wp.Kernel,
@@ -1376,18 +1388,15 @@ def _launch_integrate_kernel(
         if not add_to_output:
             output.zero_()
 
-        def as_2d_array(array):
-            return wp.array(
-                data=None,
-                ptr=array.ptr,
-                capacity=array.capacity,
-                device=array.device,
+        output_view = (
+            output
+            if output.ndim == 2
+            else _as_2d_array(
+                output,
                 shape=(test.space_partition.node_count(), test.node_dof_count),
                 dtype=type_scalar_type(output_dtype),
-                grad=None if array.grad is None else as_2d_array(array.grad),
             )
-
-        output_view = output if output.ndim == 2 else as_2d_array(output)
+        )
 
         if nodal:
             wp.launch(
@@ -1484,9 +1493,9 @@ def _launch_integrate_kernel(
     else:
         nnz = test.space_restriction.total_node_element_count() * trial.space.topology.MAX_NODES_PER_ELEMENT
 
-    triplet_rows_temp = cache.borrow_temporary(temporary_store, shape=(nnz,), dtype=int, device=device)
-    triplet_cols_temp = cache.borrow_temporary(temporary_store, shape=(nnz,), dtype=int, device=device)
-    triplet_values_temp = cache.borrow_temporary(
+    triplet_rows = cache.borrow_temporary(temporary_store, shape=(nnz,), dtype=int, device=device)
+    triplet_cols = cache.borrow_temporary(temporary_store, shape=(nnz,), dtype=int, device=device)
+    triplet_values = cache.borrow_temporary(
         temporary_store,
         shape=(
             nnz,
@@ -1496,9 +1505,6 @@ def _launch_integrate_kernel(
         dtype=output_dtype,
         device=device,
     )
-    triplet_cols = triplet_cols_temp.array
-    triplet_rows = triplet_rows_temp.array
-    triplet_values = triplet_values_temp.array
 
     if nodal:
         wp.launch(
@@ -1636,9 +1642,9 @@ def _launch_integrate_kernel(
     bsr_set_from_triplets(bsr_result, triplet_rows, triplet_cols, triplet_values, **(bsr_options or {}))
 
     # Do not wait for garbage collection
-    triplet_values_temp.release()
-    triplet_rows_temp.release()
-    triplet_cols_temp.release()
+    triplet_values.release()
+    triplet_rows.release()
+    triplet_cols.release()
 
     if add_to_output:
         output += bsr_result
@@ -2359,17 +2365,14 @@ def _launch_interpolate_kernel(
     if dest.block_shape[1] != trial.node_dof_count:
         raise RuntimeError(f"'dest' matrix blocks must have {trial.node_dof_count} columns")
 
-    triplet_rows_temp = cache.borrow_temporary(temporary_store, shape=(nnz,), dtype=int, device=device)
-    triplet_cols_temp = cache.borrow_temporary(temporary_store, shape=(nnz,), dtype=int, device=device)
-    triplet_values_temp = cache.borrow_temporary(
+    triplet_rows = cache.borrow_temporary(temporary_store, shape=(nnz,), dtype=int, device=device)
+    triplet_cols = cache.borrow_temporary(temporary_store, shape=(nnz,), dtype=int, device=device)
+    triplet_values = cache.borrow_temporary(
         temporary_store,
         dtype=dest.scalar_type,
         shape=(nnz, *dest.block_shape),
         device=device,
     )
-    triplet_cols = triplet_cols_temp.array
-    triplet_rows = triplet_rows_temp.array
-    triplet_values = triplet_values_temp.array
     triplet_rows.fill_(-1)
 
     trial_partition_arg = trial.space_partition.partition_arg_value(device)
@@ -2395,6 +2398,10 @@ def _launch_interpolate_kernel(
     )
 
     bsr_set_from_triplets(dest, triplet_rows, triplet_cols, triplet_values, **(bsr_options or {}))
+
+    triplet_values.release()
+    triplet_rows.release()
+    triplet_cols.release()
 
 
 @integrand