+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Fix type inference errors when passing reference arguments (such as array elements) to built-in functions
   ([GH-1071](https://github.com/NVIDIA/warp/issues/1071)).
 - Improve documentation and error messages about requiring a BVH for `fem.lookup` and related functionality ([GH-1072](https://github.com/NVIDIA/warp/issues/1072)).
+- Fix reference cycles introduced by `warp.fem.Temporary` and `warp.fem.ShapeBasisSpace` ([GH-1076](https://github.com/NVIDIA/warp/issues/1076)).
 
 ## [1.10.0] - 2025-11-02